Tentative deploy to github.io
Commit adds a tentative deployment to a github.io page which does not
exist at the current moment through means of a github actions.
The YML specification is relatively simple. The only manual component is
that the sphinx version is specified "hard-coded" within the action file
to ensure that the requirements can then be fully installed dynamically.